refactor api key section
