+++ apps/admin_panel/assets/src/omg-page-api/ClientKeySection.js
@@ -1,0 +1,207 @@
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
+import styled from 'styled-components'
+import Table from '../omg-table'
+import { Switch, Icon } from '../omg-uikit'
+import ApiKeysFetcher from '../omg-api-keys/apiKeysFetcher'
+import { connect } from 'react-redux'
+import { compose } from 'recompose'
+import { createApiKey, updateApiKey } from '../omg-api-keys/action'
+import CreateClientKeyModal from '../omg-create-client-key-modal'
+import queryString from 'query-string'
+import { withRouter } from 'react-router-dom'
+import moment from 'moment'
+import Copy from '../omg-copy'
+const KeySection = styled.div`
+  position: relative;
+  p {
+    color: ${props => props.theme.colors.B100};
+  }
+  h3,
+  p {
+    max-width: 800px;
+  }
+  td {
+    color: ${props => props.theme.colors.B200};
+  }
+  h3 {
+    margin-bottom: 20px;
+  }
+  tr:hover {
+    td:nth-child(1) {
+      i {
+        visibility: visible;
+      }
+    }
+  }
+  td {
+    white-space: nowrap;
+  }
+  td:nth-child(1) {
+    width: 50%;
+    border: none;
+    position: relative;
+    :before {
+      content: '';
+      position: absolute;
+      right: 0;
+      bottom: -1px;
+      height: 1px;
+      width: calc(100% - 50px);
+      border-bottom: 1px solid ${props => props.theme.colors.S200};
+    }
+  }
+  i[name='Copy'] {
+    cursor: pointer;
+    visibility: hidden;
+    cursor: pointer;
+    color: ${props => props.theme.colors.S500};
+    :hover {
+      color: ${props => props.theme.colors.B300};
+    }
+  }
+`
+
+const KeyContainer = styled.div`
+  white-space: nowrap;
+  span {
+    vertical-align: middle;
+  }
+
+  i {
+    margin-right: 5px;
+  }
+  i[name='Key'] {
+    margin-right: 5px;
+    color: ${props => props.theme.colors.B100};
+    padding: 8px;
+    border-radius: 6px;
+    border: 1px solid ${props => props.theme.colors.S400};
+  }
+  i[name='People'] {
+    color: inherit;
+  }
+`
+
+const columnsApiKey = [
+  { key: 'key', title: 'CLIENT KEY' },
+  { key: 'created_at', title: 'CREATED DATE' },
+  { key: 'status', title: 'STATUS' }
+]
+const enhance = compose(
+  withRouter,
+  connect(
+    null,
+    { createApiKey, updateApiKey }
+  )
+)
+class ApiKeyPage extends Component {
+  static propTypes = {
+    createApiKey: PropTypes.func,
+    updateApiKey: PropTypes.func,
+    location: PropTypes.object,
+    createClientKeyModalOpen: PropTypes.bool,
+    onRequestClose: PropTypes.func
+  }
+  state = {
+    createClientKeyModalOpen: false
+  }
+  onRequestClose = () => {
+    this.setState({
+      createClientKeyModalOpen: false
+    })
+  }
+
+  onClickCreateClientKey = fetch => async e => {
+    this.setState({ submitStatus: 'SUBMITTING' })
+    try {
+      await this.props.createApiKey({ name: this.state.name })
+      fetch()
+      this.onRequestClose()
+      this.setState({ submitStatus: 'SUCCESS' })
+    } catch (error) {
+      this.setState({ submitStatus: 'FAILED' })
+    }
+  }
+  onClickSwitch = ({ id, expired, fetch }) => async e => {
+    await this.props.updateApiKey({ id, expired })
+  }
+
+  rowApiKeyRenderer = fetch => (key, data, rows) => {
+    switch (key) {
+      case 'status':
+        return (
+          <Switch
+            open={!data}
+            onClick={this.onClickSwitch({ id: rows.id, expired: !rows.status, fetch })}
+          />
+        )
+      case 'key':
+        return (
+          <KeyContainer>
+            <Icon name='Key' /> <span>{data}</span> <Copy data={data} />
+          </KeyContainer>
+        )
+      case 'user':
+        return (
+          <KeyContainer>
+            <Icon name='Profile' /> <span>{data}</span>
+          </KeyContainer>
+        )
+      case 'created_at':
+        return moment(data).format()
+      default:
+        return data
+    }
+  }
+
+  render () {
+    return (
+      <ApiKeysFetcher
+        query={{
+          page: queryString.parse(this.props.location.search)['api_key_page'],
+          perPage: 10
+        }}
+        render={({ data, individualLoadingStatus, pagination, fetch }) => {
+          const apiKeysRows = data
+            .filter(key => !key.deleted_at)
+            .map(key => {
+              return {
+                key: key.key,
+                id: key.id,
+                user: key.account_id,
+                created_at: key.created_at,
+                status: key.expired,
+                updated_at: key.updated_at
+              }
+            })
+          return (
+            <KeySection>
+              <Table
+                loadingRowNumber={6}
+                rows={apiKeysRows}
+                rowRenderer={this.rowApiKeyRenderer(fetch)}
+                columns={columnsApiKey}
+                perPage={10}
+                loadingColNumber={4}
+                loadingStatus={individualLoadingStatus}
+                navigation
+                pageEntity='api_key_page'
+                isFirstPage={pagination.is_first_page}
+                isLastPage={pagination.is_last_page}
+              />
+              <CreateClientKeyModal
+                open={this.props.createClientKeyModalOpen}
+                onRequestClose={this.props.onRequestClose}
+                onOk={this.onClickCreateClientKey(fetch)}
+                loading={this.state.submitStatus === 'SUBMITTING'}
+              />
+            </KeySection>
+          )
+        }}
+      />
+    )
+  }
+}
+
+export default enhance(ApiKeyPage)
