+++ apps/admin_panel/assets/src/omg-page-api/index.js
@@ -2,133 +2,16 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import TopNavigation from '../omg-page-layout/TopNavigation'
-import { Button, Switch, Icon } from '../omg-uikit'
-import Table from '../omg-table'
-import ApiKeysFetcher from '../omg-api-keys/apiKeysFetcher'
-import AccessKeyFetcher from '../omg-access-key/accessKeysFetcher'
-import moment from 'moment'
-import ConfirmationModal from '../omg-confirmation-modal'
-import { connect } from 'react-redux'
+import { Button } from '../omg-uikit'
 import { compose } from 'recompose'
-import { createApiKey, updateApiKey } from '../omg-api-keys/action'
-import { createAccessKey, updateAccessKey } from '../omg-access-key/action'
-import CreateAdminKeyModal from '../omg-create-admin-key-modal'
-import queryString from 'query-string'
+import AdminKeySection from './AdminKeySection'
 import { withRouter, Link } from 'react-router-dom'
-import Copy from '../omg-copy'
+import ClientKeySection from './ClientKeySection'
 
 const ApiKeyContainer = styled.div`
   padding-bottom: 50px;
 `
-const KeySection = styled.div`
-  position: relative;
-  p {
-    color: ${props => props.theme.colors.B100};
-  }
-  h3,
-  p {
-    max-width: 800px;
-  }
-  td {
-    color: ${props => props.theme.colors.B200};
-  }
-  h3 {
-    margin-bottom: 20px;
-  }
-  tr:hover {
-    td:nth-child(1) {
-      i {
-        visibility: visible;
-      }
-    }
-  }
-  td {
-    white-space: nowrap;
-  }
-  td:nth-child(1) {
-    width: 50%;
-    border: none;
-    position: relative;
-    :before {
-      content: '';
-      position: absolute;
-      right: 0;
-      bottom: -1px;
-      height: 1px;
-      width: calc(100% - 50px);
-      border-bottom: 1px solid ${props => props.theme.colors.S200};
-    }
-  }
-  i[name='Copy'] {
-    cursor: pointer;
-    visibility: hidden;
-    cursor: pointer;
-    color: ${props => props.theme.colors.S500};
-    :hover {
-      color: ${props => props.theme.colors.B300};
-    }
-  }
-`
-const ConfirmCreateKeyContainer = styled.div`
-  font-size: 16px;
-  padding: 30px;
-  h4 {
-    padding-bottom: 10px;
-  }
-  p {
-    font-size: 12px;
-    max-width: 350px;
-    margin-bottom: 10px;
-  }
-  input {
-    border: 1px solid #1a56f0;
-    border-radius: 2px;
-    background-color: #ffffff;
-    width: 370px;
-    padding: 5px;
-    margin-top: 5px;
-    margin-right: 5px;
-    color: ${props => props.theme.colors.B300};
-  }
-  i[name='Copy'] {
-    margin-left: 5px;
-    cursor: pointer;
-    color: ${props => props.theme.colors.S500};
-    :hover {
-      color: ${props => props.theme.colors.B300};
-    }
-  }
-`
-const KeyContainer = styled.div`
-  white-space: nowrap;
-  span {
-    vertical-align: middle;
-  }
 
-  i {
-    margin-right: 5px;
-  }
-  i[name='Key'] {
-    margin-right: 5px;
-    color: ${props => props.theme.colors.B100};
-    padding: 8px;
-    border-radius: 6px;
-    border: 1px solid ${props => props.theme.colors.S400};
-  }
-  i[name='People'] {
-    color: inherit;
-  }
-`
-
-const InputContainer = styled.div`
-  :not(:last-child) {
-    margin-bottom: 10px;
-  }
-`
-const InputLabel = styled.div`
-  font-size: 14px;
-  color: ${props => props.theme.colors.B100};
-`
 const KeyTopBar = styled.div`
   margin-bottom: 20px;
   p {
@@ -157,70 +40,30 @@ const KeyButton = styled.button`
 const KeyTopButtonsContainer = styled.div`
   margin: 25px 0;
 `
-const columnsApiKey = [
-  { key: 'key', title: 'CLIENT KEY' },
-  { key: 'created_at', title: 'CREATED DATE' },
-  { key: 'status', title: 'STATUS' }
-]
-const columnsAdminKeys = [
-  { key: 'key', title: 'ADMIN KEY' },
-  { key: 'name', title: 'NAME' },
-  { key: 'created_at', title: 'CREATED DATE' },
-  { key: 'status', title: 'STATUS' },
-  { key: 'global_role', title: 'GLOBAL ROLE' }
-]
-const enhance = compose(
-  withRouter,
-  connect(
-    null,
-    { createApiKey, updateApiKey, createAccessKey, updateAccessKey }
-  )
-)
+const enhance = compose(withRouter)
 class ApiKeyPage extends Component {
   static propTypes = {
-    createApiKey: PropTypes.func,
-    createAccessKey: PropTypes.func,
-    updateApiKey: PropTypes.func,
     divider: PropTypes.bool,
-    location: PropTypes.object,
-    updateAccessKey: PropTypes.func,
     match: PropTypes.object
   }
   state = {
-    accessModalOpen: false,
-    ewalletModalOpen: false
+    createAdminKeyModalOpen: false,
+    createClientKeyModalOpen: false
   }
   onRequestClose = () => {
     this.setState({
-      accessModalOpen: false,
-      ewalletModalOpen: false
+      createAdminKeyModalOpen: false,
+      createClientKeyModalOpen: false
     })
   }
-  onRequestCloseShowPrivateKey = () => {
-    this.setState({
-      secretKey: '',
-      accessKey: '',
-      submitStatus: 'DEFAULT',
-      privateKeyModalOpen: false
-    })
+
+  onClickCreateClientKey = e => {
+    this.setState({ createClientKeyModalOpen: true })
   }
-  onClickCreateAccessKey = e => {
-    this.setState({ accessModalOpen: true })
+  onClickCreateAdminKey = e => {
+    this.setState({ createAdminKeyModalOpen: true })
   }
-  onClickCreateEwalletKey = e => {
-    this.setState({ ewalletModalOpen: true })
-  }
-  onClickOkCreateEwalletKey = fetch => async e => {
-    this.setState({ submitStatus: 'SUBMITTING' })
-    try {
-      await this.props.createApiKey()
-      fetch()
-      this.onRequestClose()
-      this.setState({ submitStatus: 'SUCCESS' })
-    } catch (error) {
-      this.setState({ submitStatus: 'FAILED' })
-    }
-  }
+
   onSubmitSuccess = fetch => data => {
     fetch()
     this.setState({
@@ -228,190 +71,6 @@ class ApiKeyPage extends Component {
       accessKey: data.access_key,
       privateKeyModalOpen: true
     })
-  }
-  onClickSwitch = ({ id, expired, fetch }) => async e => {
-    await this.props.updateApiKey({ id, expired })
-  }
-  onClickAccessKeySwitch = ({ id, expired, fetch }) => async e => {
-    await this.props.updateAccessKey({ id, expired })
-  }
-
-  rowApiKeyRenderer = fetch => (key, data, rows) => {
-    switch (key) {
-      case 'status':
-        return (
-          <Switch
-            open={!data}
-            onClick={this.onClickSwitch({ id: rows.id, expired: !rows.status, fetch })}
-          />
-        )
-      case 'key':
-        return (
-          <KeyContainer>
-            <Icon name='Key' /> <span>{data}</span> <Copy data={data} />
-          </KeyContainer>
-        )
-      case 'user':
-        return (
-          <KeyContainer>
-            <Icon name='Profile' /> <span>{data}</span>
-          </KeyContainer>
-        )
-      case 'created_at':
-        return moment(data).format()
-      default:
-        return data
-    }
-  }
-
-  rowAccessKeyRenderer = fetch => (key, data, rows) => {
-    switch (key) {
-      case 'status':
-        return (
-          <Switch
-            open={!data}
-            onClick={this.onClickAccessKeySwitch({ id: rows.id, expired: !rows.status, fetch })}
-          />
-        )
-      case 'key':
-        return (
-          <KeyContainer>
-            <Icon name='Key' /> <span>{data}</span> <Copy data={data} />
-          </KeyContainer>
-        )
-      case 'user':
-        return (
-          <KeyContainer>
-            <Icon name='Profile' /> <span>{data}</span>
-          </KeyContainer>
-        )
-      case 'created_at':
-        return moment(data).format()
-      default:
-        return data
-    }
-  }
-  renderClientKey () {
-    return (
-      <ApiKeysFetcher
-        query={{
-          page: queryString.parse(this.props.location.search)['api_key_page'],
-          perPage: 10
-        }}
-        render={({ data, individualLoadingStatus, pagination, fetch }) => {
-          const apiKeysRows = data
-            .filter(key => !key.deleted_at)
-            .map(key => {
-              return {
-                key: key.key,
-                id: key.id,
-                user: key.account_id,
-                created_at: key.created_at,
-                status: key.expired,
-                updated_at: key.updated_at
-              }
-            })
-          return (
-            <KeySection>
-              <Table
-                loadingRowNumber={6}
-                rows={apiKeysRows}
-                rowRenderer={this.rowApiKeyRenderer(fetch)}
-                columns={columnsApiKey}
-                perPage={10}
-                loadingColNumber={4}
-                loadingStatus={individualLoadingStatus}
-                navigation
-                pageEntity='api_key_page'
-                isFirstPage={pagination.is_first_page}
-                isLastPage={pagination.is_last_page}
-              />
-              <ConfirmationModal
-                open={this.state.ewalletModalOpen}
-                onRequestClose={this.onRequestClose}
-                onOk={this.onClickOkCreateEwalletKey(fetch)}
-                loading={this.state.submitStatus === 'SUBMITTING'}
-              >
-                <ConfirmCreateKeyContainer>
-                  <h4>Generate Client Key</h4>
-                  <p>Are you sure you want to generate eWallet key ?</p>
-                </ConfirmCreateKeyContainer>
-              </ConfirmationModal>
-            </KeySection>
-          )
-        }}
-      />
-    )
-  }
-  renderAdminKey = () => {
-    return (
-      <AccessKeyFetcher
-        query={{
-          page: queryString.parse(this.props.location.search)['access_key_page'],
-          perPage: 10
-        }}
-        render={({ data, individualLoadingStatus, pagination, fetch }) => {
-          const apiKeysRows = data.map(key => {
-            return {
-              key: key.access_key,
-              id: key.id,
-              user: key.account_id,
-              created_at: key.created_at,
-              status: key.expired,
-              name: key.name || 'Not Provided',
-              global_role: key.global_role || 'none'
-            }
-          })
-          return (
-            <KeySection>
-              <Table
-                loadingRowNumber={6}
-                rows={apiKeysRows}
-                rowRenderer={this.rowAccessKeyRenderer(fetch)}
-                columns={columnsAdminKeys}
-                loadingStatus={individualLoadingStatus}
-                navigation
-                isFirstPage={pagination.is_first_page}
-                isLastPage={pagination.is_last_page}
-                pageEntity='access_key_page'
-              />
-              <CreateAdminKeyModal
-                open={this.state.accessModalOpen}
-                onRequestClose={this.onRequestClose}
-                onSubmitSuccess={this.onSubmitSuccess(fetch)}
-                closeTimeoutMS={0}
-                loading={this.state.submitStatus === 'SUBMITTING'}
-              />
-              <ConfirmationModal
-                open={this.state.privateKeyModalOpen}
-                onRequestClose={this.onRequestCloseShowPrivateKey}
-                onOk={this.onRequestCloseShowPrivateKey}
-                confirmText='Got it!'
-                cancel={false}
-              >
-                <ConfirmCreateKeyContainer>
-                  <h4>Your key pair</h4>
-                  <p>
-                    Please copy and keep this pair of acesss and secret key. Secret key will use to
-                    open your encrypted information.
-                  </p>
-                  <InputContainer>
-                    <InputLabel>Access Key</InputLabel>
-                    <input value={this.state.accessKey} spellCheck='false' />
-                    <Copy data={this.state.accessKey} />
-                  </InputContainer>
-                  <InputContainer>
-                    <InputLabel>Secret key</InputLabel>
-                    <input value={this.state.secretKey} spellCheck='false' />
-                    <Copy data={this.state.secretKey} />
-                  </InputContainer>
-                </ConfirmCreateKeyContainer>
-              </ConfirmationModal>
-            </KeySection>
-          )
-        }}
-      />
-    )
   }
 
   render () {
@@ -437,11 +96,11 @@ class ApiKeyPage extends Component {
               <KeyButton active={activeTab === 'client'}>Client Keys</KeyButton>
             </Link>
             {activeTab === 'admin' ? (
-              <Button size='small' onClick={this.onClickCreateAccessKey} styleType={'secondary'}>
+              <Button size='small' onClick={this.onClickCreateAdminKey} styleType={'secondary'}>
                 <span>Generate Admin Key</span>
               </Button>
             ) : (
-              <Button size='small' onClick={this.onClickCreateEwalletKey} styleType={'secondary'}>
+              <Button size='small' onClick={this.onClickCreateClientKey} styleType={'secondary'}>
                 <span>Generate Client Key</span>
               </Button>
             )}
@@ -459,7 +118,17 @@ class ApiKeyPage extends Component {
             </p>
           )}
         </KeyTopBar>
-        {activeTab === 'admin' ? this.renderAdminKey() : this.renderClientKey()}
+        {activeTab === 'admin' ? (
+          <AdminKeySection
+            createAdminKeyModalOpen={this.state.createAdminKeyModalOpen}
+            onRequestClose={this.onRequestClose}
+          />
+        ) : (
+          <ClientKeySection
+            createClientKeyModalOpen={this.state.createClientKeyModalOpen}
+            onRequestClose={this.onRequestClose}
+          />
+        )}
       </ApiKeyContainer>
     )
   }
